+++ src/__init__.py
@@ -1,4 +1,4 @@
 from .env_conf import EnvKeyNotFound, EnvValueUnexpected, get_credentials
 from driver_selector import get_webdriver
-from page_object import PageObject
+from page_objects import PageObject, LOGIN_PAGE_URL
 from message_activities import MessageActivities
