+++ src/page_objects/__init__.py
@@ -1,0 +1,2 @@
+from .page_object import PageObject
+from .consts import LOGIN_PAGE_URL
