+++ src/page_objects/consts.py
@@ -1,0 +1,8 @@
+
+LOGIN_PAGE = "LOGIN_PAGE-state"
+LOGIN_PAGE_URL = "https://account.mail.ru/login?success_redirect=https://octavius.mail.ru&allow_external=1"
+
+
+STATES = {
+    LOGIN_PAGE_URL: LOGIN_PAGE
+}
