+++ src/page_objects/states.py
@@ -1,0 +1,6 @@
+from consts import *
+
+def get_state(url):
+    if url in STATES:
+        return STATES[url]
+    return None
