+++ tests/__init__.py
@@ -1,4 +1,4 @@
 from src import *
 
 from .applying_filters import TestApplyingFilters
-import src
+
